requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
selenium==4.15.2
python-dotenv==1.0.0
flask==3.0.0
